Split logic into services, more tests

+++ Assets/Scripts/Components/SellItemEvent.cs
@@ -1,6 +1,6 @@
 namespace EconomicsGame.Components {
 	public struct SellItemEvent {
-		public long Count;
-		public long PricePerUnit;
+		public double Count;
+		public double PricePerUnit;
 	}
 }
